--- config/initializers/requirements.rb
@@ -1,1 +1,0 @@
-require Rails.root.join('app/models/user_week_calendar.rb')
